raspberry_pi/requirements.txt
google-cloud-storage
pyudev
